Schramski's Algorithm Updated.
Remove the annotation recording that you made changes and when. Add this annotation via a git commit.
Try to manually wrap comments so that they can be more easily read when they are viewed in non-wrapped views. You can do this by hitting return after a sentence reaches about half the width of your window.